# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SaveYourMoney API
  version: "1.0.0"
servers:
  - url: http://127.0.0.1:5000
...
paths:
...
  /gastos:
    get:
      tags: [Expenses]
      summary: Lista gastos
      parameters:
        - in: query
          name: mes
          schema: { type: integer }
        - in: query
          name: ano
          schema: { type: integer }
        - in: query
          name: recorrente
          schema: { type: string, enum: [todos, sim, nao] }
      responses:
        "200":
          description: Lista de gastos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      tags: [Expenses]
      summary: Cria gasto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseCreate"
      responses:
        "201":
          description: Gasto criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /gastos/{id}:
    put:
      tags: [Expenses]
      summary: Atualiza gasto
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Gasto atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Gasto não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Expenses]
      summary: Exclui gasto
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                scope:
                  type: string
                  enum: [this, future, all]
      responses:
        "200":
          description: Gasto excluído
        "404":
          description: Gasto não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /entradas:
    get:
      tags: [Incomes]
      summary: Lista entradas
      parameters:
        - in: query
          name: mes
          schema: { type: integer }
        - in: query
          name: ano
          schema: { type: integer }
      responses:
        "200":
          description: Lista de entradas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Income"
    post:
      tags: [Incomes]
      summary: Cria entrada
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeCreate"
      responses:
        "201":
          description: Entrada criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
        "400":
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /entradas/{id}:
    put:
      tags: [Incomes]
      summary: Atualiza entrada
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeUpdate"
      responses:
        "200":
          description: Entrada atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
        "404":
          description: Entrada não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Incomes]
      summary: Exclui entrada
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Entrada excluída
        "404":
          description: Entrada não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /parcelas:
    get:
      tags: [Installments]
      summary: Lista parcelas
      parameters:
        - in: query
          name: cartao_id
          schema: { type: integer }
        - in: query
          name: card_id
          schema: { type: integer }
        - in: query
          name: mes
          schema: { type: integer }
        - in: query
          name: ano
          schema: { type: integer }
      responses:
        "200":
          description: Lista de parcelas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Installment"
  /faturas:
    get:
      tags: [Installments]
      summary: Calcula fatura
      parameters:
        - in: query
          name: cartao_id
          schema: { type: integer }
        - in: query
          name: card_id
          schema: { type: integer }
        - in: query
          name: mes
          schema: { type: integer }
        - in: query
          name: ano
          schema: { type: integer }
      responses:
        "200":
          description: Fatura do cartão
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: number }
                  parcelas:
                    type: array
                    items:
                      $ref: "#/components/schemas/Installment"
                required: [total, parcelas]
        "400":
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /metas:
    get:
      tags: [Goals]
      summary: Lista metas
      parameters:
        - in: query
          name: mes
          schema: { type: integer }
        - in: query
          name: ano
          schema: { type: integer }
      responses:
        "200":
          description: Lista de metas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      tags: [Goals]
      summary: Cria meta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalCreate"
      responses:
        "201":
          description: Meta criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "400":
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /metas/{id}:
    put:
      tags: [Goals]
      summary: Atualiza meta
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Meta atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Meta não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Goals]
      summary: Exclui meta
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Meta excluída
        "404":
          description: Meta não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Expense:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        value: { type: number }
        month: { type: integer }
        year: { type: integer }
        category_id: { type: integer, nullable: true }
        recurrence_id: { type: integer, nullable: true }
        payment_method: { type: string }
        notes: { type: string, nullable: true }
      required: [id, name, value, month, year]
    ExpenseCreate:
      type: object
      properties:
        name: { type: string }
        value: { type: number }
        month: { type: integer }
        year: { type: integer }
        category_id: { type: integer, nullable: true }
        payment_method: { type: string }
        notes: { type: string, nullable: true }
        installments:
          type: object
          properties:
            card_id: { type: integer }
            total: { type: integer }
          required: [card_id, total]
        recurring:
          type: object
          properties:
            enabled: { type: boolean }
            frequency: { type: string }
            interval_months: { type: integer }
            occurrences: { type: integer }
            end_month: { type: integer }
            end_year: { type: integer }
      required: [name, value, month, year]
    ExpenseUpdate:
      type: object
      properties:
        name: { type: string }
        value: { type: number }
        month: { type: integer }
        year: { type: integer }
        category_id: { type: integer, nullable: true }
        payment_method: { type: string }
        notes: { type: string, nullable: true }
        scope: { type: string, enum: [this, future] }
      required: [name, value, month, year]
    Income:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        value: { type: number }
        month: { type: integer }
        year: { type: integer }
        confirmed: { type: boolean }
        notes: { type: string, nullable: true }
      required: [id, name, value, month, year, confirmed]
    IncomeCreate:
      type: object
      properties:
        name: { type: string }
        value: { type: number }
        month: { type: integer }
        year: { type: integer }
        confirmed: { type: boolean }
        notes: { type: string, nullable: true }
      required: [name, value, month, year]
    IncomeUpdate:
      type: object
      properties:
        name: { type: string }
        value: { type: number }
        month: { type: integer }
        year: { type: integer }
        confirmed: { type: boolean }
        notes: { type: string, nullable: true }
      required: [name, value, month, year]
...
    Installment:
      type: object
      properties:
        id: { type: integer }
        card_id: { type: integer }
        expense_name: { type: string }
        installment_number: { type: integer }
        total_installments: { type: integer }
        value: { type: number }
        month: { type: integer }
        year: { type: integer }
        status: { type: string }
      required: [card_id, expense_name, installment_number, total_installments, value, month, year]
...
    Goal:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        limit_value: { type: number }
        month: { type: integer }
        year: { type: integer }
        category_id: { type: integer, nullable: true }
      required: [id, name, limit_value, month, year]
    GoalCreate:
      type: object
      properties:
        name: { type: string }
        limit_value: { type: number }
        month: { type: integer }
        year: { type: integer }
        category_id: { type: integer, nullable: true }
      required: [name, limit_value, month, year]
    GoalUpdate:
      type: object
      properties:
        name: { type: string }
        limit_value: { type: number }
        month: { type: integer }
        year: { type: integer }
        category_id: { type: integer, nullable: true }
      required: [name, limit_value, month, year]
...